requirements.txt
Flask>=1.0,<=1.1.2
newsapi-python==0.2.6
numpy==1.19.5
matplotlib==3.3.4
openai==0.3.0
scikit-learn==0.24.1
requests==2.25.0
umap-learn==0.5.1
webhoseio==0.5